% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculate_mean.R
\name{calculate_mean}
\alias{calculate_mean}
\title{Title Batch Caculate Mean}
\usage{
calculate_mean(data, y_variable, x_variables, output_file)
}
\arguments{
\item{data}{data which you need to call.}

\item{y_variable}{outcome variable.}

\item{x_variables}{grouping variable.}

\item{output_file}{The test result will be output as an excel file. This
parameter requires you to write a file name with the suffix "xlsx", such as:
"outcome.xlsx"}
}
\description{
This function provides a simple and efficient way to calculate the mean value
of one or more variables in a dataset. It allows users to compute the means for
multiple variables simultaneously, making it convenient for analyzing and
summarizing data.
}
\examples{
my_data <- data.frame(
Group = c("A", "A", "B", "B"),
Value1 = c(10, 15, 12, 18),
Value2 = c(5, 8, 6, 9))
calculate_mean(data = my_data, y_variable = "Value1", x_variables = c("Group"),
output_file = "output.xlsx")
}
