% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculate_sd.R
\name{calculate_sd}
\alias{calculate_sd}
\title{Title Batch Caculate SD}
\usage{
calculate_sd(data, y_variable, x_variables, output_file)
}
\arguments{
\item{data}{data which you need to call.}

\item{y_variable}{outcome variable.}

\item{x_variables}{grouping variable.}

\item{output_file}{The test result will be output as an excel file. This
parameter requires you to write a file name with the suffix "xlsx", such as:
"outcome.xlsx"}
}
\description{
This function provides a function to calculate the standard
deviation of one or more variables in a dataset and save the results to an
Excel file using the openxlsx package.
}
\examples{
my_data <- data.frame(
Group = c("A", "A", "B", "B"),
Value1 = c(10, 15, 12, 18),
Value2 = c(5, 8, 6, 9))
calculate_sd(data = my_data, y_variable = "Value1", x_variables = c("Group"),
output_file = "output.xlsx")
}
